// packages/ui/src/types/sanity.ts
export interface SanityImage {
  _type: 'image';
  asset: {
    _id?: string;
    _ref: string;
    _type: 'reference';
    url?: string;
    metadata?: {
      dimensions?: {
        aspectRatio?: number;
        height?: number;
        width?: number;
      };
      lqip?: string;
      blurHash?: string;
    }
  };
  crop?: {
    _type: 'sanity.imageCrop';
    bottom: number;
    left: number;
    right: number;
    top: number;
  };
  hotspot?: {
    _type: 'sanity.imageHotspot';
    height: number;
    width: number;
    x: number;
    y: number;
  };
  alt?: string;
}

// New rich image type used in hero and other sections
export interface SanityRichImage {
  _type: 'richImage';
  _key?: string;
  alt: string; // required in schema
  image: SanityImage; // nested Sanity image object
}

export interface SanityVideo {
  _type: 'video';
  title?: string;
  videoType?: string; 
  youtubeUrl?: string;
  wistiaUrl?: string;
  videoFile?: {
    asset?: {url?: string};
  };
  thumbnail?: SanityImage;
}